requirements.txt
google-generativeai
flask 
jinja2
python-dotenv
chromadb